Clarify correspondence relations of the internal lists
To clarify correspondence relations of the argument of local-function,
its initial value, and internal temporal variables.

lambda-list-info = (mapcar #'list lambda-list gs-lambda-list gs
init-params)

lambda-list is a list of arguments of user defined local-function.
init-params is a list of actual parameters for user defined
local-function.
gs-lambda-list and gs are lists of temporal gensym variables.